FIXES #23!
This took way longer than it should have, but it was a complex bug spanning many files. Basically, it is related to the fact that tiles weren't being registered, so they wouldn't be spawned in, so the tileget  packet sorter would throw a nullpointerexeption, which would be caught and errored as a bad packet sorter warning.